requirements.txt
certifi==2018.1.18
chardet==3.0.4
idna==2.6
msgpack-python==0.5.6
python-memcached==1.59
redis==2.10.6
requests==2.18.4
six==1.11.0
urllib3==1.22
